Create traits for the CMSObject error and property handling
